requirements.txt
astroid==2.3.3
autopep8==1.5
cycler==0.10.0
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.18.1
pandas==1.0.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
PyQt5==5.14.0
PyQt5-sip==12.7.1
pyqtgraph==0.10.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.15
typed-ast==1.4.1
wrapt==1.11.2
xlrd==1.2.0
